+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `ResourceBase`
+  - Increased code coverage.
+
 ## [1.0.0] - 2022-12-31
 
 ### Added