+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `ResourceBase`
+  - Fixed style changed.
+
 ## [1.1.0] - 2023-02-26
 
 ### Added